fixed an issue in the station where it askes for the location in meters and tries to detect an int while it gets a double also added the function the calculate x,y from the stations in the server the solution is almost ready. the missing part is added a checking for every input from the user

+++ src/sami/alabri/company/Main.java
@@ -54,7 +54,7 @@ public class Main {
 
             if(!isCalibrated) {
                 System.out.println("enter the distance of the device from the station in meters ");
-                int distanceFromDeviceInMeters = sc.nextInt();
+                double distanceFromDeviceInMeters = sc.nextDouble();
 
                 System.out.println("now the wifi equation to calculate the distance will be calibrated ");
 
@@ -97,7 +97,7 @@ public class Main {
         clnt.sendData(Double.toString(length));
     }
 
-    private static void calibration(int distance){
+    private static void calibration(double distance){
 
 
         for (int i = 0; i < 12; i++) {
